fix #18: minor code refractor and typo fixes
