+++ unitex/src/fr/umlv/unitex/frames/GraphFrame.java
@@ -82,7 +82,6 @@ import fr.umlv.unitex.config.Preferences;
 import fr.umlv.unitex.config.PreferencesManager;
 import fr.umlv.unitex.diff.GraphDecorator;
 import fr.umlv.unitex.files.FileUtil;
-import fr.umlv.unitex.findandreplace.FindAndReplaceData;
 import fr.umlv.unitex.graphrendering.GenericGraphBox;
 import fr.umlv.unitex.graphrendering.GraphBox;
 import fr.umlv.unitex.graphrendering.GraphicalZone;
