+++ unitex/src/fr/umlv/unitex/frames/InternalFrameManager.java
@@ -38,7 +38,6 @@ import fr.umlv.unitex.config.ConfigManager;
 import fr.umlv.unitex.config.PreferencesManager;
 import fr.umlv.unitex.diff.GraphDecorator;
 import fr.umlv.unitex.exceptions.UserRefusedFrameClosingError;
-import fr.umlv.unitex.findandreplace.FindAndReplaceData;
 import fr.umlv.unitex.grf.GraphPresentationInfo;
 import fr.umlv.unitex.io.GraphIO;
 import fr.umlv.unitex.listeners.DelaFrameListener;
