+++ unitex/src/fr/umlv/unitex/graphtools/FindAndReplace.java
@@ -18,7 +18,7 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA.
  *
  */
-package fr.umlv.unitex.findandreplace;
+package fr.umlv.unitex.graphtools;
 
 import fr.umlv.unitex.graphrendering.GenericGraphBox;
 import fr.umlv.unitex.graphrendering.GraphBox;
